Lab 4, Exercise 13/14
Finally figured out why general protection faults kept randomly
appearing when running with multiple CPUs...external interrupts are
usually turned off, but seem to be turned back on when a CPU halts (so
that a timer can wake it back up).  This sets up the IDT to handle
hardware interrupts, so that the CPU trap generator doesn't turn them
into GPFs.